fix: change credits type from int to float in PrivyAuthResponse schemaUpdated the credits field in the PrivyAuthResponse model to use a float type instead of an int, allowing for more precise credit representation.

+++ src/schemas/auth.py
@@ -54,5 +54,5 @@ class PrivyAuthResponse(BaseModel):
     is_new_user: Optional[bool] = None
     display_name: Optional[str] = None
     email: Optional[str] = None
-    credits: Optional[int] = None
+    credits: Optional[float] = None
     timestamp: Optional[datetime] = None
